+++ packages/blocks/core/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added support for the `Blocksy Widgets Wrapper Block` by Blockera.
 - Added support for the `Blocksy About Me Block` by Blockera.
 - Added support for the `Blocksy About Me Block` text inner block by Blockera.
+- Added support for the `Blocksy About Me Block` icons inner block by Blockera.
 
 
 ### Automated Tests